% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characters.R
\name{characters}
\alias{characters}
\title{characters}
\usage{
characters(name = FALSE, id = FALSE)
}
\arguments{
\item{name}{Name of the title, fuzzy name is supported. Default as `FALSE`.}

\item{id}{The id of your inquiry title. Default as `FALSE`.}
}
\value{
Names of characters, as a dataframe.
}
\description{
This function gets the names of roles and actors/actresses of a movie/TV show.
}
\examples{

# Gets the characters of "game of thrones".
top_crew(name = "game of thrones")

# Gets the characters of id="tt0944947".
top_crew(id = "tt0944947")
}
\author{
Yunong Li
}
