% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_id.R
\name{get_id}
\alias{get_id}
\title{get_id}
\usage{
get_id(name)
}
\arguments{
\item{name}{Name of the title, fuzzy name is supported.}
}
\value{
The id of the title.
}
\description{
This funtion gets the imdb id for a certain title.
}
\examples{

# Gets the id of "game of thrones".
get_id("game of thrones")

}
\author{
Yunong Li
}
