% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/more.R
\name{more}
\alias{more}
\title{more}
\usage{
more(name = FALSE, id = FALSE)
}
\arguments{
\item{name}{Name of the title, fuzzy name is supported. Default as `FALSE`.}

\item{id}{The id of your inquiry title. Default as `FALSE`.}
}
\value{
Basic information of similar titles, as a dataframe.
}
\description{
This function gets more similar movies/TV shows of the input.
}
\examples{

# Gets the characters of "game of thrones".
more(name = "game of thrones")

# Gets the characters of id="tt0944947".
more(id = "tt0944947")
}
\author{
Yunong Li
}
