% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overview.R
\name{overview}
\alias{overview}
\title{overview}
\usage{
overview(name = FALSE, id = FALSE)
}
\arguments{
\item{name}{Name of the title, fuzzy name is supported. Default as `FALSE`.}

\item{id}{The id of your inquiry title. Default as `FALSE`.}
}
\value{
The overview of your inquiry title, as a dataframe.
}
\description{
This funtion gets the overview for a certain title.
}
\examples{

# Gets the overview of "game of thrones".
overview(name = "game of thrones")

# Gets the overview of id="tt0944947".
overview(id = "tt0944947")
}
\author{
Yunong Li
}
