% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_crew.R
\name{top_crew}
\alias{top_crew}
\title{top_crew}
\usage{
top_crew(name = FALSE, id = FALSE)
}
\arguments{
\item{name}{Name of the title, fuzzy name is supported. Default as `FALSE`.}

\item{id}{The id of your inquiry title. Default as `FALSE`.}
}
\value{
Names of top crew, as a dataframe.
}
\description{
This function gets the names of directors and writers of a movie/TV show.
}
\examples{

# Gets the overview of "game of thrones".
top_crew(name = "game of thrones")

# Gets the overview of id="tt0944947".
top_crew(id = "tt0944947")
}
\author{
Yunong Li
}
